requirements.txt
blinker==1.4
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Mail==0.9.1
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
python-dotenv==0.19.2
Werkzeug==2.0.3
